Updated ip packets to use unsigned values

+++ lib/src/main/java/network/grape/lib/network/ip/IpHeader.java
@@ -6,12 +6,12 @@ import java.net.InetAddress;
  * Abstraction of common functions to both Ip4 and Ip6 headers.
  */
 public interface IpHeader {
-  byte getProtocol();
+  short getProtocol();
   InetAddress getSourceAddress();
   InetAddress getDestinationAddress();
 
-  byte IP4_VERSION = 0x04;
-  byte IP6_VERSION = 0x06;
+  short IP4_VERSION = 0x04;
+  short IP6_VERSION = 0x06;
   int IP4HEADER_LEN = 20;
   int IP4_WORD_LEN = 4;
   int IP6HEADER_LEN = 40;
